Remove s3 modules
